refactor: 리뷰 저장 url 주소 변경 및 controller 변경
